+++ Scripts/SettingsHandler.gd
@@ -9,17 +9,19 @@ static var image_default: bool = false
 # godot does not have a Set datatype. lame!
 static var opened_channels: Array = []
 static var websocket_address: String = Main.DSCR_URL
+static var theme_color: int = 57
 
 static func initialize() -> void:
 	# Setting names kept in PascalCase to keep backwards compatibility
 	# also for some reason TMfDS uses PascalCase too. blegh.
-	do_formatting = SaveSystem.settings.get_or_add("DoFormatting", true)
-	image_default = SaveSystem.settings.get_or_add("ImageDefault", false)
-	truncate_message_size = SaveSystem.settings.get_or_add("TruncateMessageSize", 63)
-	preferred_callsign = SaveSystem.settings.get_or_add("PreferredCallsign", -1)
-	font_size = SaveSystem.settings.get_or_add("FontSize", 18)
+	do_formatting = SaveSystem.settings.get_or_add("DoFormatting", do_formatting)
+	image_default = SaveSystem.settings.get_or_add("ImageDefault", image_default)
+	truncate_message_size = SaveSystem.settings.get_or_add("TruncateMessageSize", truncate_message_size)
+	preferred_callsign = SaveSystem.settings.get_or_add("PreferredCallsign", preferred_callsign)
+	font_size = SaveSystem.settings.get_or_add("FontSize", font_size)
 	opened_channels = SaveSystem.settings.get_or_add("OpenedChannels", []).map(func (a): return int(a))
 	websocket_address = SaveSystem.settings.get_or_add("WebsocketAddress", Main.DSCR_URL)
+	theme_color = SaveSystem.settings.get_or_add("ThemeColor", theme_color)
 
 static func save() -> void:
 	SaveSystem.save_settings()
@@ -32,3 +34,4 @@ static func export() -> void:
 	SaveSystem.settings.set("FontSize", font_size)
 	SaveSystem.settings.set("OpenedChannels", opened_channels)
 	SaveSystem.settings.set("WebsocketAddress", websocket_address)
+	SaveSystem.settings.set("ThemeColor", theme_color)