+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added autocomplete! Hell yeah
 * Added icon
 * Settings tab now has an option to change websocket and/or attempt manual reconnect
+* Added theme color setting
 
 ### Changes